Changed Expr to Tokens.
Expr only represents the vector of Op (tokens). However, functions
should be represented as the vector of vector of tokens as each parameter
should be represented as vector of tokens also. (e.g f(1+1, 2+2))

To introduce the concept of function, we have to create a new struct
that can have the vector of vector of tokens. I will use "Expr" for this
instead which captures the meaning more properly.